requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.6.2.post1
appdirs==1.4.4
appnope==0.1.4
asttokens==2.4.1
asyncio==3.4.3
attrs==24.2.0
azure-ai-documentintelligence==1.0.0b4
azure-core==1.32.0
backoff==2.2.1
beanie==1.27.0
beautifulsoup4==4.12.3
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.3.1
cryptography==43.0.3
cycler==0.12.1
dataclasses-json==0.6.7
datasets==3.1.0
debugpy==1.8.8
decorator==5.1.1
deepdiff==8.0.1
Deprecated==1.2.14
dill==0.3.8
distro==1.9.0
dnspython==2.7.0
effdet==0.4.1
email_validator==2.2.0
emoji==2.14.0
executing==2.1.0
fastapi==0.115.5
fastapi-cli==0.0.5
filelock==3.16.1
filetype==1.2.0
flatbuffers==24.3.25
fonttools==4.54.1
frozenlist==1.5.0
fsspec==2024.9.0
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.23.0
google-auth==2.36.0
google-cloud-vision==3.8.0
googleapis-common-protos==1.66.0
greenlet==3.1.1
gritql==0.1.5
grpcio==1.67.1
grpcio-status==1.67.1
h11==0.14.0
# hdbscan==0.8.39
html5lib==1.1
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.26.2
humanfriendly==10.0
idna==3.10
iopath==0.1.10
ipykernel==6.29.5
ipython==8.29.0
isodate==0.7.2
jedi==0.19.2
Jinja2==3.1.4
jiter==0.7.1
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
langchain==0.3.7
langchain-community==0.3.7
langchain-core==0.3.17
langchain-experimental==0.3.3
langchain-openai==0.2.8
langchain-text-splitters==0.3.2
langchain-unstructured==0.1.5
langdetect==1.0.9
langgraph==0.2.46
langgraph-checkpoint==2.0.3
langgraph-sdk==0.1.35
langserve==0.3.0
langsmith==0.1.142
layoutparser==0.3.4
lazy-model==0.2.0
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
memory-profiler==0.61.0
motor==3.6.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
neo4j==5.26.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
olefile==0.47
omegaconf==2.3.0
onnx==1.17.0
onnxruntime==1.20.0
openai==1.54.4
opencv-python==4.10.0.84
orderly-set==5.2.2
orjson==3.10.11
packaging==24.2
pandas==2.2.3
parso==0.8.4
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.4
pexpect==4.9.0
pi_heif==0.20.0
pikepdf==9.4.0
pillow==11.0.0
platformdirs==4.3.6
portalocker==2.10.1
prompt_toolkit==3.0.48
propcache==0.2.0
proto-plus==1.25.0
protobuf==5.28.3
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==18.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycocotools==2.0.8
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
Pygments==2.18.0
pymongo==4.9.2
pyparsing==3.2.0
pypdf==5.1.0
pypdfium2==4.30.0
pysbd==0.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-iso639==2024.10.22
python-magic==0.4.27
python-multipart==0.0.17
python-oxmsg==0.0.1
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
ragas==0.2.6
RapidFuzz==3.10.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rsa==4.9
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.35
sse-starlette==1.8.2
stack-data==0.6.3
starlette==0.41.2
sympy==1.13.3
tenacity==9.0.0
threadpoolctl==3.5.0
tiktoken==0.8.0
timm==1.0.11
tokenizers==0.20.3
toml==0.10.2
tomlkit==0.12.5
torch==2.2.2
torchvision==0.17.2
tornado==6.4.1
tqdm==4.67.0
traitlets==5.14.3
transformers==4.46.2
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
unstructured==0.16.5
unstructured-client==0.25.9
unstructured-inference==0.8.1
unstructured.pytesseract==0.3.13
urllib3==2.2.3
uvicorn==0.23.2
watchfiles==0.24.0
wcwidth==0.2.13
webencodings==0.5.1
websockets==14.0
wrapt==1.16.0
xxhash==3.5.0
yarl==1.17.1
